+++ ui/App.vue
@@ -6,6 +6,7 @@
         </main>
         <main-footer></main-footer>
         <audio id="sharp" src="/sounds/sharp.ogg" crossOrigin="anonymous"></audio>
+        <app-update-available></app-update-available>
         <communication-error></communication-error>
         <dataset-import-failure></dataset-import-failure>
     </div>