+++ ui/app.ts
@@ -19,6 +19,7 @@ import MemoryLevel from './components/MemoryLevel.vue';
 import MemoryUsageChart from './components/MemoryUsageChart.vue';
 import ProcessInfo from './components/ProcessInfo.vue';
 import ServerSettings from './components/ServerSettings.vue';
+import AppUpdateAvailable from './components/AppUpdateAvailable.vue';
 import CommunicationError from './components/CommunicationError.vue';
 import DatasetImportFailure from './components/DatasetImportFailure.vue';
 import { ValidationProvider, ValidationObserver, extend } from 'vee-validate';
@@ -28,6 +29,7 @@ import { createHttpLink } from 'apollo-link-http';
 import { InMemoryCache } from 'apollo-cache-inmemory';
 import { Workbox } from 'workbox-window';
 import routes from './routes';
+import bus from  './bus';
 
 require('./scss/app.scss');
 
@@ -37,6 +39,12 @@ require('./scss/app.scss');
 
 if ('serviceWorker' in navigator) {
     const wb = new Workbox('/sw.js');
+
+    wb.addEventListener('waiting', (event) => {
+        if (event.isUpdate) {
+            bus.$emit('service-worker-installed');
+        }
+    });
   
     wb.register();
 }
@@ -77,6 +85,7 @@ Vue.component('memory-level', MemoryLevel);
 Vue.component('memory-usage-chart', MemoryUsageChart);
 Vue.component('process-info', ProcessInfo);
 Vue.component('server-settings', ServerSettings);
+Vue.component('app-update-available', AppUpdateAvailable);
 Vue.component('communication-error', CommunicationError);
 Vue.component('dataset-import-failure', DatasetImportFailure);