+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - 0.2.6-beta
     - Added export chart feature to Web UI
+    - Add web app update notification
 
 - 0.2.5-beta
     - Added Backoff and Retry client middleware